requirements.txt
Django==1.8
argparse==1.2.1
django-tastypie==0.12.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
wsgiref==0.1.2
